Address Clippy warnings

+++ src/fish.rs
@@ -496,10 +496,10 @@ fn write_subword_wrapper_fn<W: Write>(
     let max_fallback_level = dfa.get_max_fallback_level().unwrap_or(ARRAY_START as usize);
 
     let mut completion_literals: Vec<HashMap<StateId, Vec<usize>>> =
-        vec![Default::default(); max_fallback_level + 1 as usize];
+        vec![Default::default(); max_fallback_level + 1_usize];
 
     let mut completion_commands: Vec<HashMap<StateId, Vec<usize>>> =
-        vec![Default::default(); max_fallback_level + 1 as usize];
+        vec![Default::default(); max_fallback_level + 1_usize];
 
     for (from, input_id, _) in dfa.iter_transitions() {
         match dfa.get_input(input_id).clone() {
